% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_classicalmds.R
\name{bclassicalmds}
\alias{bclassicalmds}
\title{Classical Multidimensional Scaling with Similarity/Dissimilarity Input}
\usage{
bclassicalmds(
  x,
  type,
  dim = 2,
  add = FALSE,
  plot = TRUE,
  labels = NULL,
  jitter = 0
)
}
\arguments{
\item{x}{A square matrix of similarities or dissimilarities}

\item{type}{Character string specifying the input type. Must be one of
"similarities", "dissimilarities", or any unambiguous abbreviation
(e.g., "s", "sim", "d", "dis", "diss"). If missing, the user will be
prompted to specify.}

\item{dim}{Number of dimensions for the MDS solution (default = 2)}

\item{add}{Logical, whether to add a constant to make eigenvalues non-negative (default = FALSE)}

\item{plot}{Logical, whether to produce a plot (default = TRUE)}

\item{labels}{Optional character vector of labels for points}

\item{jitter}{Numeric, proportion of the range to jitter plotted points by (default = 0, no jitter)}
}
\value{
An object of class "bmds" containing:
\item{points}{Matrix of MDS coordinates}
\item{eig}{Eigenvalues}
\item{GOF}{Goodness of fit statistics}
\item{stress}{Stress measure}
}
\description{
Performs classical multidimensional scaling on a similarity or dissimilarity matrix
and optionally produces a plot of the results.
}
\examples{
# Using dissimilarity matrix
bclassicalmds(as.matrix(dist(USArrests)), "d", dim=2)
bclassicalmds(as.matrix(dist(USArrests)), "dissimilarities", dim=2)

# Using similarity matrix (correlation)
cor_mat <- cor(mtcars)
bclassicalmds(cor_mat, "s", dim=2)
bclassicalmds(cor_mat, "similarities", dim=3, plot=FALSE)

# With jitter to reduce overplotting
bclassicalmds(cor_mat, "s", dim=2, jitter=0.02)

}
