% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_utils.R
\name{bclip}
\alias{bclip}
\alias{bsnip}
\title{Copy data fram to the clipboard}
\usage{
bclip(df)

bsnip(df)
}
\arguments{
\item{df}{A data frame or tibble.}
}
\value{
data.frame with columns: variable, type
}
\description{
Copy data fram to the clipboard
}
\examples{
bclip(iris)
}
