% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_corresp.R
\name{bcorresp.contrib.plot}
\alias{bcorresp.contrib.plot}
\title{Create a contribution plot for bcorresp results}
\usage{
bcorresp.contrib.plot(
  ca_result,
  dimension = 1,
  type = c("row", "column"),
  n_display = 10,
  main = NULL
)
}
\arguments{
\item{ca_result}{A bcorresp object}

\item{dimension}{Which dimension to plot contributions for (default = 1)}

\item{type}{Type of contributions to plot: "row" or "column"}

\item{n_display}{Number of top contributors to display (default = 10)}

\item{main}{Main title for the plot (optional)}
}
\description{
Create a contribution plot for bcorresp results
}
