% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_corresp.R
\name{bcorresp.coord}
\alias{bcorresp.coord}
\title{Extract coordinates from bcorresp results}
\usage{
bcorresp.coord(
  ca_result,
  type = c("row", "column", "both"),
  dimensions = c(1, 2)
)
}
\arguments{
\item{ca_result}{A bcorresp object}

\item{type}{Type of coordinates to extract: "row", "column", or "both"}

\item{dimensions}{Which dimensions to extract (default = c(1, 2))}
}
\description{
Extract coordinates from bcorresp results
}
