% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_corresp.R
\name{bcorresp.export}
\alias{bcorresp.export}
\title{Export bcorresp results to CSV files}
\usage{
bcorresp.export(
  ca_result,
  file_prefix = "bcorresp_results",
  include_data = FALSE
)
}
\arguments{
\item{ca_result}{A bcorresp object}

\item{file_prefix}{Prefix for exported file names (default = "bcorresp_results")}

\item{include_data}{Whether to include the original data (default = FALSE)}
}
\description{
Export bcorresp results to CSV files
}
