% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_corresp.R
\name{bcorresp.interpret}
\alias{bcorresp.interpret}
\title{Interpret bcorresp results with guidance}
\usage{
bcorresp.interpret(ca_result, dimensions = c(1, 2), n_points = 5)
}
\arguments{
\item{ca_result}{A bcorresp object}

\item{dimensions}{Which dimensions to interpret (default = c(1, 2))}

\item{n_points}{Number of points to highlight (default = 5)}
}
\description{
Interpret bcorresp results with guidance
}
