% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_readgoogle.R
\name{bgoogle}
\alias{bgoogle}
\title{Short alias for breadgoogle}
\usage{
bgoogle(...)
}
\arguments{
\item{...}{All arguments passed to breadgoogle}
}
\value{
A data frame
}
\description{
Short alias for breadgoogle
}
