% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_lda.R
\name{bldaterms}
\alias{bldaterms}
\title{Extract top terms from blda object}
\usage{
bldaterms(x, n = 10)
}
\arguments{
\item{x}{A blda object}

\item{n}{Number of top terms per topic. Default is 10.}
}
\value{
A matrix with terms as rows and topics as columns
}
\description{
Extract top terms from blda object
}
