% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_stats.R
\name{bmean}
\alias{bmean}
\title{Mean with na.rm = TRUE by default}
\usage{
bmean(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Additional arguments passed to \code{\link[base:mean]{base::mean()}}.}
}
\value{
Numeric mean of x.
}
\description{
Wrapper around base::mean() that always removes NAs unless overridden.
}
