% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borgworld-package.R
\docType{package}
\name{borgworld-package}
\alias{borgworld}
\alias{borgworld-package}
\title{borgworld: Simple statistical and data processing utilities}
\description{
All functions are prefixed with \code{b}.
}
\section{Getting started}{

\code{help(package = "borgworld")} lists all topics.
\code{browseVignettes("borgworld")} opens tutorials.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StephenBorgatti/borgworld}
  \item Report bugs at \url{https://github.com/StephenBorgatti/borgworld/issues}
}

}
\author{
\strong{Maintainer}: Steve Borgatti \email{sborgatti@uky.edu}

}
\keyword{internal}
