% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_paste.R
\name{bpaste}
\alias{bpaste}
\title{Read a tabular block from the clipboard}
\usage{
bpaste(header = TRUE, sep = "\\t", auto_rownames = TRUE)
}
\arguments{
\item{header}{Logical. First row has names. Default TRUE.}

\item{sep}{Field separator. Default tab.}

\item{auto_rownames}{Logical. Automatically convert first column to row names
when it's non-numeric or has no header. Default TRUE.}
}
\value{
data.frame
}
\description{
Read a tabular block from the clipboard
}
\examples{
\dontrun{
# Copy a range in Excel, then:
bpaste()
}
}
