% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_stats.R
\name{bsd}
\alias{bsd}
\title{Column SDs for numeric columns, long format
These are "population sds" using the n divisor}
\usage{
bsd(df)
}
\arguments{
\item{df}{A data frame or tibble.}
}
\value{
Base data.frame with columns: variable, sd.
}
\description{
Column SDs for numeric columns, long format
These are "population sds" using the n divisor
}
