% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_data.R
\name{bstandardize}
\alias{bstandardize}
\title{Standardize columns of df}
\usage{
bstandardize(df, k = 3)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame of z-scores
Non-finite inputs return NA.
}
\description{
Standardize columns of df
}
\examples{
  # Simple scatter
  bstandardize(hdi)
}
