% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_stats.R
\name{bsum}
\alias{bsum}
\title{Sum with na.rm = TRUE by default}
\usage{
bsum(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Additional arguments passed to \code{\link[base:sum]{base::sum()}}.}
}
\value{
Numeric sum of x.
}
\description{
Wrapper around base::sum() that always removes NAs unless overridden.
}
