% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_stats.R
\name{bsummarize}
\alias{bsummarize}
\title{Summarize numeric columns}
\usage{
bsummarize(df)
}
\arguments{
\item{df}{data.frame or tibble}
}
\value{
data.frame with columns: variable, min, max, median, mean, sd, n_missing
}
\description{
Summarize numeric columns
}
