% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_utils.R
\name{bvars}
\alias{bvars}
\title{List variables and their data types}
\usage{
bvars(df)
}
\arguments{
\item{df}{A data frame or tibble.}
}
\value{
data.frame with columns: variable, type
}
\description{
List variables and their data types
}
\examples{
bvars(iris)
}
