% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_compare_mds.R
\name{compare_mds}
\alias{compare_mds}
\title{Compare Multiple MDS Methods}
\usage{
compare_mds(x, type, dim = 2, plot = TRUE, ...)
}
\arguments{
\item{x}{Input matrix (similarity or dissimilarity)}

\item{type}{Character string specifying the input type. Must be one of
"similarities", "dissimilarities", or any unambiguous abbreviation.
If missing, the user will be prompted to specify.}

\item{dim}{Number of dimensions (default = 2)}

\item{plot}{Logical, whether to create comparison plots (default = TRUE)}

\item{...}{Additional arguments passed to MDS functions}
}
\value{
A list of class "mds_comparison" containing both MDS results and comparison statistics
}
\description{
Runs both classical and non-metric MDS on the same data for comparison.
}
\examples{
# Compare methods on dissimilarity matrix
dist_mat <- as.matrix(dist(USArrests))
comparison <- compare_mds(dist_mat, "d")

# Compare methods on correlation matrix
cor_mat <- cor(mtcars)
comparison <- compare_mds(cor_mat, "similarities", dim = 3)

}
