% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_lda.R
\name{predict.blda}
\alias{predict.blda}
\title{Get topic assignments for new documents}
\usage{
\method{predict}{blda}(object, newdata, text_col, ...)
}
\arguments{
\item{object}{A blda object (containing the fitted model)}

\item{newdata}{A data frame with new documents}

\item{text_col}{The name or index of the text column}

\item{...}{Additional arguments (ignored)}
}
\value{
A data frame with topic distributions for new documents
}
\description{
Get topic assignments for new documents
}
