% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_corresp.R
\name{print.bcorresp}
\alias{print.bcorresp}
\title{Print method for bcorresp objects}
\usage{
\method{print}{bcorresp}(x, ...)
}
\arguments{
\item{x}{A bcorresp object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for bcorresp objects
}
