% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_eigenvalues.R
\name{print.beigenvalues}
\alias{print.beigenvalues}
\title{Print method for beigenvalues}
\usage{
\method{print}{beigenvalues}(x, digits = 4, n = NULL, ...)
}
\description{
Print method for beigenvalues
}
