% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_hiclus.R
\name{print.bhiclus}
\alias{print.bhiclus}
\title{Print method for bhiclus objects}
\usage{
\method{print}{bhiclus}(x, ...)
}
\arguments{
\item{x}{A bhiclus object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for bhiclus objects
}
