% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_lda.R
\name{print.blda}
\alias{print.blda}
\title{Print method for blda objects}
\usage{
\method{print}{blda}(x, n_terms = 10, n_docs = 6, ...)
}
\arguments{
\item{x}{A blda object}

\item{n_terms}{Number of top terms to display per topic. Default is 10.}

\item{n_docs}{Number of document rows to preview. Default is 6.}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for blda objects
}
