% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_mds.R
\name{print.bmds}
\alias{print.bmds}
\title{Print method for bmds objects}
\usage{
\method{print}{bmds}(x, ...)
}
\arguments{
\item{x}{A bmds object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for bmds objects
}
