% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_nonmetricmds.R
\name{print.bnmds}
\alias{print.bnmds}
\title{Print method for bnmds objects}
\usage{
\method{print}{bnmds}(x, ...)
}
\arguments{
\item{x}{A bnmds object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for bnmds objects
}
