% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_pca.R
\name{print.bpca}
\alias{print.bpca}
\title{Print method for bpca objects}
\usage{
\method{print}{bpca}(x, ...)
}
\arguments{
\item{x}{A bpca object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for bpca objects
}
