% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_compare_mds.R
\name{print.mds_comparison}
\alias{print.mds_comparison}
\title{Print Method for MDS Comparison}
\usage{
\method{print}{mds_comparison}(x, ...)
}
\arguments{
\item{x}{An mds_comparison object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for MDS Comparison
}
