% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_nonmetricmds.R
\name{shepard.bnmds}
\alias{shepard.bnmds}
\title{Shepard plot for bnmds objects}
\usage{
shepard.bnmds(x, original_dist = NULL, ...)
}
\arguments{
\item{x}{A bnmds object}

\item{original_dist}{Optional original dissimilarity matrix. If not provided,
the function will not be able to create the plot.}

\item{...}{Additional arguments passed to plot()}
}
\description{
Creates a Shepard diagram to assess the fit of non-metric MDS
}
