% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_corresp.R
\name{summary.bcorresp}
\alias{summary.bcorresp}
\title{Summary method for bcorresp objects}
\usage{
\method{summary}{bcorresp}(object, n_display = 5, ...)
}
\arguments{
\item{object}{A bcorresp object}

\item{n_display}{Number of items to display (default = 5)}

\item{...}{Additional arguments (not used)}
}
\description{
Summary method for bcorresp objects
}
