% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_hiclus.R
\name{summary.bhiclus}
\alias{summary.bhiclus}
\title{Summary method for bhiclus objects}
\usage{
\method{summary}{bhiclus}(object, ...)
}
\arguments{
\item{object}{A bhiclus object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for bhiclus objects
}
