% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_mds.R
\name{summary.bmds.unified}
\alias{summary.bmds.unified}
\title{Summary Method for Unified bmds Objects}
\usage{
\method{summary}{bmds.unified}(object, ...)
}
\arguments{
\item{object}{A bmds.unified object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the object
}
\description{
Summary Method for Unified bmds Objects
}
