% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_pca.R
\name{summary.bpca}
\alias{summary.bpca}
\title{Summary method for bpca objects}
\usage{
\method{summary}{bpca}(object, ...)
}
\arguments{
\item{object}{A bpca object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for bpca objects
}
