% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_compare_mds.R
\name{summary.mds_comparison}
\alias{summary.mds_comparison}
\title{Summary Method for MDS Comparison}
\usage{
\method{summary}{mds_comparison}(object, ...)
}
\arguments{
\item{object}{An mds_comparison object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary Method for MDS Comparison
}
