% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_utils.R
\name{wd}
\alias{wd}
\title{get working directroy}
\usage{
wd()
}
\value{
working direcotry
}
\description{
Accepts one or more paths (e.g., pasted from Explorer) and returns
normalized paths with your chosen slash.
}
\examples{
wd()
}
