% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_classicalmds.R
\name{bclassicalmds}
\alias{bclassicalmds}
\title{Classical Multidimensional Scaling with Similarity/Dissimilarity Input}
\usage{
bclassicalmds(
  x,
  type = c("s", "d"),
  k = 2,
  add = FALSE,
  plot = TRUE,
  labels = NULL
)
}
\arguments{
\item{x}{A square matrix of similarities or dissimilarities}

\item{type}{Character string, either "s" for similarity or "d" for dissimilarity}

\item{k}{Number of dimensions for the MDS solution (default = 2)}

\item{add}{Logical, whether to add a constant to make eigenvalues non-negative (default = FALSE)}

\item{plot}{Logical, whether to produce a plot (default = TRUE)}

\item{labels}{Optional character vector of labels for points}
}
\value{
An object of class "bmds" containing:
\item{points}{Matrix of MDS coordinates}
\item{eig}{Eigenvalues}
\item{GOF}{Goodness of fit statistics}
\item{stress}{Stress measure}
}
\description{
Performs classical multidimensional scaling on a similarity or dissimilarity matrix
and optionally produces a plot of the results.
}
\examples{
# Using dissimilarity matrix
bclassicalmds(as.matrix(dist(USArrests)), type="d", k=2)

# Using similarity matrix (convert correlation to similarity)
cor_mat <- cor(mtcars)
bclassicalmds(cor_mat, type="s", k=2)

}
