% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_utils.R
\name{bpaste}
\alias{bpaste}
\title{Read a tabular block from the clipboard}
\usage{
bpaste(header = TRUE, sep = "\\t")
}
\arguments{
\item{header}{Logical. First row has names. Default TRUE.}

\item{sep}{Field separator. Default tab.}
}
\value{
data.frame
}
\description{
Read a tabular block from the clipboard
}
\examples{
\dontrun{
# Copy a range in Excel, then:
bpaste()
}
}
